+++ app/src/main/java/comschakravorti21/github/cubeassist/CubeView.java
@@ -6,6 +6,8 @@ import android.graphics.Color;
 import android.graphics.Paint;
 import android.os.AsyncTask;
 import android.support.annotation.Nullable;
+import android.support.v4.content.ContextCompat;
+import android.support.v7.app.AppCompatActivity;
 import android.util.AttributeSet;
 import android.util.TypedValue;
 import android.view.MotionEvent;
@@ -15,8 +17,6 @@ import android.widget.TextView;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
-
-import static comschakravorti21.github.cubeassist.Cube.generateRandScramble;
 
 /**
  * Created by development on 7/28/17.
@@ -32,6 +32,8 @@ public class CubeView extends View {
     private int speedMultiplier;
 
     private Cube cube = new Cube();
+    private boolean hasColorInput;
+    private char[][][] colorsInputted;
     //Default scramble
     private final String DEFAULT_SCRAMBLE = getResources().getString(R.string.default_scramble);
     private String scramble = DEFAULT_SCRAMBLE,
@@ -102,31 +104,33 @@ public class CubeView extends View {
 
     private void init() {
         cube = new Cube();
+        moveSet = new String[8];
         resetScramble(scramble);
 
+        Context context = getContext();
         bluePaint = new Paint(Paint.ANTI_ALIAS_FLAG);
         bluePaint.setStyle(Paint.Style.FILL);
-        bluePaint.setColor(Color.parseColor("#03A9F4"));
+        bluePaint.setColor(ContextCompat.getColor(context, R.color.cubeBlue));
 
         greenPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
         greenPaint.setStyle(Paint.Style.FILL);
-        greenPaint.setColor(Color.parseColor("#FF11CF31"));
+        greenPaint.setColor(ContextCompat.getColor(context, R.color.cubeGreen));
 
         redPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
         redPaint.setStyle(Paint.Style.FILL);
-        redPaint.setColor(Color.parseColor("#FFE70000"));
+        redPaint.setColor(ContextCompat.getColor(context, R.color.cubeRed));
 
         orangePaint = new Paint(Paint.ANTI_ALIAS_FLAG);
         orangePaint.setStyle(Paint.Style.FILL);
-        orangePaint.setColor(Color.parseColor("#FFFE7D15"));
+        orangePaint.setColor(ContextCompat.getColor(context, R.color.cubeOrange));
 
         whitePaint = new Paint(Paint.ANTI_ALIAS_FLAG);
         whitePaint.setStyle(Paint.Style.FILL);
-        whitePaint.setColor(Color.parseColor("#FFFEFEFE"));
+        whitePaint.setColor(ContextCompat.getColor(context, R.color.cubeWhite));
 
         yellowPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
         yellowPaint.setStyle(Paint.Style.FILL);
-        yellowPaint.setColor(Color.parseColor("#FFFCD51E"));
+        yellowPaint.setColor(ContextCompat.getColor(context, R.color.cubeYellow));
 
         strokePaint = new Paint(Paint.ANTI_ALIAS_FLAG);
         strokePaint.setStyle(Paint.Style.STROKE);
@@ -409,25 +413,79 @@ public class CubeView extends View {
         movesIndex = 0; phase = 0;
         phaseString = "Sunflower";
         invalidate();
+        hasColorInput = false;
 
-        moveSet = new String[] {sunflower, whiteCross, whiteCorners,
-                secondLayer, yellowCross, OLL, PLL, " "};
+        moveSet[0] = sunflower;
+        moveSet[1] = whiteCross;
+        moveSet[2] = whiteCorners;
+        moveSet[3] = secondLayer;
+        moveSet[4] = yellowCross;
+        moveSet[5] = OLL;
+        moveSet[6] = PLL;
+        moveSet[7] = " ";
+
 
         UpdateUI updateMoves = new UpdateUI(UpdateUI.UPDATE_MOVES_ON_UI);
         updateMoves.execute(getContext());
     }
 
-    public void resetCurrentScramble() {
+    /**
+     * After the user inputs their desired colors in color selection mode, pressing the setInputs button
+     * will invoke this method, acquiring the required moves necessary to solve the cube. The cube is restored back to
+     * the scrambled state after the solution moves are acquired.
+     */
+    public void resetScrambleByColorInputs(char[][][] colorsInputted) {
+        this.colorsInputted = colorsInputted;
         cube = new Cube();
-        cube.scramble(scramble);
+        cube.setAllColors(colorsInputted);
+
+        sunflower = cube.makeSunflower();
+        whiteCross = cube.makeWhiteCross();
+        whiteCorners = cube.finishWhiteLayer();
+        secondLayer = cube.insertAllEdges();
+        yellowCross = cube.makeYellowCross();
+        OLL = cube.orientLastLayer();
+        PLL = cube.permuteLastLayer();
+
+        cube = new Cube();
+        cube.setAllColors(colorsInputted);
+        //If the cube is being scrambled newly after initializing is complete and animation has begun,
+        //be sure to reset all reference indexes
+        movesIndex = 0; phase = 0;
+        phaseString = "Sunflower";
+        hasColorInput = true;
+        scramble = "";
+
+        moveSet[0] = sunflower;
+        moveSet[1] = whiteCross;
+        moveSet[2] = whiteCorners;
+        moveSet[3] = secondLayer;
+        moveSet[4] = yellowCross;
+        moveSet[5] = OLL;
+        moveSet[6] = PLL;
+        moveSet[7] = " ";
+
+        resetCurrentScramble(); //I'm not sure why, but this is necessary for it to work
+    }
+
+    public void resetCurrentScramble() {
+        if(!hasColorInput) {
+            cube = new Cube();
+            cube.scramble(scramble);
+
+        } else {
+            cube = new Cube();
+            cube.setAllColors(colorsInputted);
+        }
 
         movesToPerform = sunflower;
         movesPerformed = "";
 
-        movesIndex = 0; phase = 0;
+        movesIndex = 0;
+        phase = 0;
         phaseString = "Sunflower";
-        invalidate();
 
+        invalidate();
         UpdateUI updateMoves = new UpdateUI(UpdateUI.UPDATE_MOVES_ON_UI);
         updateMoves.execute(getContext());
     }
@@ -562,6 +620,7 @@ public class CubeView extends View {
 
                     UpdateUI updateMoves = new UpdateUI(UpdateUI.UPDATE_MOVES_ON_UI);
                     updateMoves.execute(getContext());
+
                 }
             };
 
@@ -592,6 +651,9 @@ public class CubeView extends View {
         @Override
         protected void onPreExecute() {
             super.onPreExecute();
+            if(taskType == SKIP_PHASES) {
+                stopAnimation();
+            }
         }
 
         @Override
@@ -638,8 +700,6 @@ public class CubeView extends View {
                 postInvalidate();
             }
             else if(taskType == UPDATE_MOVES_ON_UI) {
-                final SolutionActivity solutionActivity = activity;
-
                 switch(phase) {
                     case 0:
                         phaseString = "Sunflower";
@@ -667,13 +727,13 @@ public class CubeView extends View {
                         break;
                 }
 
-                activity.runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        solutionActivity.updateMoves(movesToPerform.substring(movesIndex).trim(),
-                                movesPerformed.trim(), phaseString);
-                    }
-                });
+                TextSolutionFragment fragment = (TextSolutionFragment)((AppCompatActivity)getContext())
+                        .getSupportFragmentManager().findFragmentById(R.id.container);
+
+                if(fragment != null) {
+                    fragment.updateMoves(movesToPerform.substring(movesIndex).trim(),
+                            movesPerformed.trim(), phaseString);
+                }
 
                 super.onPostExecute(activity);
             }
