+++ app/src/main/java/comschakravorti21/github/cubeassist/SolutionActivity.java
@@ -1,91 +1,86 @@
 package comschakravorti21.github.cubeassist;
 
+import android.content.res.Configuration;
+import android.support.annotation.NonNull;
+import android.support.design.widget.NavigationView;
 import android.support.v4.app.DialogFragment;
+import android.support.v4.app.Fragment;
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.SeekBar;
 import android.widget.TextView;
 
-public class SolutionActivity extends AppCompatActivity implements View.OnClickListener,
-        EditScrambleDialog.EditScrambleDialogListener, SeekBar.OnSeekBarChangeListener{
+public class SolutionActivity extends AppCompatActivity implements EditScrambleDialog.EditScrambleDialogListener{
 
-    CubeView cubeView;
+    private final String TEXT_SCRAMBLE = "text scramble";
+    private final String COLOR_INPUT = "color input";
+    private String currentMode = TEXT_SCRAMBLE;
+    private TextSolutionFragment textSolutionFragment;
+    private Toolbar toolbar;
+    private ActionBarDrawerToggle drawerToggle;
+    private DrawerLayout drawerLayout;
+    private NavigationView drawerList;
+    private String[] navDrawerTitles;
+
+    @Override
+    public void onBackPressed() {
+        if (getFragmentManager().getBackStackEntryCount() > 0 ){
+            getFragmentManager().popBackStack();
+        } else {
+            super.onBackPressed();
+        }
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_solution);
+        setContentView(R.layout.activity_test);
 
-        Toolbar toolbar = (Toolbar)findViewById(R.id.toolbar);
+        toolbar = (Toolbar)findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        cubeView = (CubeView)findViewById(R.id.cube_view);
-        findViewById(R.id.rewind).setOnClickListener(this);
-        findViewById(R.id.skip_forward).setOnClickListener(this);
-        SeekBar seekBar = (SeekBar)findViewById(R.id.speed_adjuster);
-        seekBar.setOnSeekBarChangeListener(this);
-        seekBar.setMax(14);
-    }
+        textSolutionFragment = new TextSolutionFragment();
+        getSupportFragmentManager().beginTransaction()
+                .add(R.id.container, textSolutionFragment, "Text Solution Fragment")
+                .addToBackStack(null)
+                .commit();
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_text_solution, menu);
-        return super.onCreateOptionsMenu(menu);
-    }
+        navDrawerTitles = getResources().getStringArray(R.array.nav_drawer_items);
+        drawerLayout = (DrawerLayout)findViewById(R.id.drawer_layout);
+        drawerList = (NavigationView) findViewById(R.id.left_drawer);
+        drawerList.setCheckedItem(R.id.text_scramble);
+        drawerToggle = new ActionBarDrawerToggle(this, drawerLayout,
+                toolbar, R.string.open_drawer, R.string.close_drawer) {
 
-    public void updateMoves(String movesToPerform, String movesPerformed, String phase) {
-        TextView toPerformView = (TextView)findViewById(R.id.moves_to_perform);
-        toPerformView.setText(movesToPerform);
-        TextView performedView = (TextView)findViewById(R.id.moves_performed);
-        performedView.setText(movesPerformed);
-        TextView phaseView = (TextView)findViewById(R.id.phase_view);
-        phaseView.setText(phase);
+            /** Called when a drawer has settled in a completely closed state. */
+            public void onDrawerClosed(View view) {
+                super.onDrawerClosed(view);
+            }
 
-    }
+            /** Called when a drawer has settled in a completely open state. */
+            public void onDrawerOpened(View drawerView) {
+                super.onDrawerOpened(drawerView);
+            }
+        };
 
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch(item.getItemId()) {
-            case R.id.action_random:
-                TextView scrambleView = (TextView)findViewById(R.id.scramble_view);
-                scrambleView.setText(cubeView.randScramble());
-                break;
-            case R.id.action_reset:
-                cubeView.resetCurrentScramble();
-                break;
-            case R.id.action_edit_scramble:
-                Bundle args = new Bundle();
-                args.putString(EditScrambleDialog.SCRAMBLE_TAG,
-                        ((TextView)findViewById(R.id.scramble_view)).getText().toString());
-                EditScrambleDialog dialog = new EditScrambleDialog();
-                dialog.setArguments(args);
-                dialog.show(getSupportFragmentManager(), "edit scramble");
-                break;
-        }
-        return super.onOptionsItemSelected(item);
-    }
+        drawerLayout.addDrawerListener(drawerToggle);
+        drawerList.setNavigationItemSelectedListener(new DrawerClickListener());
 
-    @Override
-    public void onClick(View view) {
-        switch (view.getId()) {
-            case R.id.skip_forward:
-                cubeView.skipToPhase(cubeView.getPhase()+1);
-                break;
-            case R.id.rewind:
-                cubeView.skipToPhase(cubeView.getPhase()-1);
-                break;
-        }
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setHomeButtonEnabled(true);
     }
 
     @Override
     public void onDialogPositiveClick(DialogFragment dialog, String scramble) {
         TextView scrambleView = (TextView)findViewById(R.id.scramble_view);
         scrambleView.setText(scramble);
-        cubeView.resetScramble(scramble);
+        textSolutionFragment.cubeView.resetScramble(scramble);
     }
 
     @Override
@@ -93,23 +88,52 @@ public class SolutionActivity extends AppCompatActivity implements View.OnClickL
 
     }
 
+    /**
+     * When using the ActionBarDrawerToggle, you must call it during
+     * onPostCreate() and onConfigurationChanged()...
+     */
     @Override
-    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-        if (cubeView.getAnimationStopped()) {
-            cubeView.setSpeedMultiplier(1 + progress); //Avoid division by 0, min is 1
-        } else {
-            cubeView.stopAnimation();
-            cubeView.startAnimation(1 + progress);
+    protected void onPostCreate(Bundle savedInstanceState) {
+        super.onPostCreate(savedInstanceState);
+        // Sync the toggle state after onRestoreInstanceState has occurred.
+        drawerToggle.syncState();
+    }
+
+    @Override
+    public void onConfigurationChanged(Configuration newConfig) {
+        super.onConfigurationChanged(newConfig);
+        // Pass any configuration change to the drawer toggle
+        drawerToggle.onConfigurationChanged(newConfig);
+    }
+
+    private class DrawerClickListener implements NavigationView.OnNavigationItemSelectedListener {
+
+        @Override
+        public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+            switch (item.getItemId()) {
+                case R.id.text_scramble:
+                    if(!currentMode.equals(TEXT_SCRAMBLE)) {
+                        getSupportFragmentManager().beginTransaction()
+                                .replace(R.id.container, new TextSolutionFragment(), "Text Solution Fragment")
+                                .addToBackStack(null)
+                                .commit();
+                        currentMode = TEXT_SCRAMBLE;
+                        drawerLayout.closeDrawers();
+                    }
+                    break;
+                case R.id.color_input:
+                    if(!currentMode.equals(COLOR_INPUT)) {
+                        getSupportFragmentManager().beginTransaction()
+                                .replace(R.id.container, new ColorInputSolutionFragment(), "Color Input Fragment")
+                                .addToBackStack(null)
+                                .commit();
+                        currentMode = COLOR_INPUT;
+                        drawerLayout.closeDrawers();
+                    }
+                    break;
+            }
+            return true;
         }
     }
 
-    @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-
-    }
-
-    @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-
-    }
 }
