+++ app/src/main/res/layout/activity_test.xml
@@ -9,7 +9,7 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="comschakravorti21.github.cubeassist.TestActivity">
+        tools:context="comschakravorti21.github.cubeassist.SolutionActivity">
 
         <android.support.v7.widget.Toolbar
             android:id="@+id/toolbar"
