+++ app/src/main/res/menu/menu_text_solution.xml
@@ -7,7 +7,7 @@
         android:id="@+id/action_reset"
         android:icon="@drawable/ic_replay_black"
         android:title="@string/reset_action"
-        app:showAsAction="always"/>
+        app:showAsAction="ifRoom"/>
 
     <item
         android:id="@+id/action_edit_scramble"
